requirements.txt
psycopg2==2.9.5
python-telegram-bot==13.11
python-dateutil==2.7.0
black==24.10.0
pre-commit==3.3.0

